requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
